--story=1002145 --user=林邦泉 流程管理&编辑页面 https://www.tapd.cn/54247054/s/1042504

+++ src/main/java/codedriver/module/process/api/form/FormDeleteApi.java
@@ -9,8 +9,10 @@ import com.alibaba.fastjson.JSONObject;
 import codedriver.framework.apiparam.core.ApiParamType;
 import codedriver.framework.process.dao.mapper.FormMapper;
 import codedriver.framework.process.exception.form.FormNotFoundException;
+import codedriver.framework.process.exception.form.FormReferencedCannotBeDeletedException;
 import codedriver.framework.restful.annotation.Description;
 import codedriver.framework.restful.annotation.Input;
+import codedriver.framework.restful.annotation.Output;
 import codedriver.framework.restful.annotation.Param;
 import codedriver.framework.restful.core.ApiComponentBase;
 
@@ -39,6 +41,9 @@ public class FormDeleteApi extends ApiComponentBase {
 	@Input({
 		@Param(name = "uuid", type = ApiParamType.STRING, isRequired = true, desc = "表单uuid")
 	})
+	@Output({
+		@Param(name = "uuid", type = ApiParamType.STRING, desc = "表单uuid")
+	})
 	@Description(desc = "表单删除接口")
 	@Override
 	public Object myDoService(JSONObject jsonObj) throws Exception {
@@ -47,11 +52,13 @@ public class FormDeleteApi extends ApiComponentBase {
 		if(formMapper.checkFormIsExists(uuid) == 0) {
 			throw new FormNotFoundException(uuid);
 		}
+		if(formMapper.getFormReferenceCount(uuid) > 0) {
+			throw new FormReferencedCannotBeDeletedException(uuid);
+		}
 		formMapper.deleteFormByUuid(uuid);
 		formMapper.deleteFormVersionByFormUuid(uuid);
 		formMapper.deleteFormAttributeByFormUuid(uuid);
-		formMapper.deleteProcessFormByFormUuid(uuid);
-		return null;
+		return uuid;
 	}
 
 }
